crypto-service-python/requirements.txt
fastapi
uvicorn
cryptography
python-dotenv
python-jose[cryptography]